server/Engine/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
huggingface-hub>=0.17.0
gtts>=2.4.0
